Refactored block model json parsing and added some more ease of functionality with the model system. If something isnt explicit, it is auto generated (only uvs, normals, cull direction, and any transformations [leave blank for no transformations])
Switched from Task.Run to ThreadPool.QueueUserWorkItem as it lifted some perf decreases for some reason. (maybe related to https://stackoverflow.com/questions/43009500/how-to-avoid-a-task-to-block-the-main-thread)

Block class has overload methods for custom functionality of blocks, which are OnBlockSet, OnBlockPlace, OnBlockMesh, and OnBlockDestroy which are currently the only methods fully implemented so far. The rest are OnBlockInteract, OnBlockMine, OnTickUpdate, and OnRandomTickUpdate.

You can now leave block models as null inside a block json if you are making a block with custom functionality. This used to not be the case where the default model would be the one specified in json. However calling base.OnBlockMesh will not do anything for custom blocks. Use ChunkMeshHandler.AddModel to add your models instead.

In order to use the properties from a block json, you need to call the SetProperties method in your custom block's constructor. You also need to set the GuiRenderableBlockModel to a model of your choice for the custom block.

Added block lighting propagation using grid based visibility. Some optimizations needed for it.

Reimplemented AO after block model refactor.

Now using sRGB textures and an hdr framebuffer.